Moved xml tests into dataExports.js test file
